Add support for ROIs generated from View's
This is a very simple implementation of something that can hopefully be improved
and cleaned up in the future: support for context-defined annotation widgets.

Each ROI is defined as a parameter, they are added to a View with an argument to
the Image_WithRoi type, which is read by the Correlator and thus creates an
ROI. When the ROI is changed, an updated parameter is sent to metropc and the
Correlator modifies the context file with the new ROI parameters (position,
height, etc).